refactor: merge devcontainer tests into main test file
- Merge all tests from sessionManager.devcontainer.test.ts into sessionManager.test.ts
- Add missing test cases for devcontainer functionality
- Remove separate devcontainer test file to consolidate all tests
- Fix test isolation issues by clearing mocks between tests
